a_action_chatbot/chatbot.py
import configparser
import google.generativeai as genai
import os

class Chatbot:
    def __init__(self):
        config = configparser.ConfigParser()
        # Get the absolute path to the directory containing this script
        script_dir = os.path.dirname(os.path.abspath(__file__))
        # Construct the absolute path to config.ini
        config_path = os.path.join(script_dir, 'config.ini')
        config.read(config_path)
        try:
            self.api_key = config['GEMINI']['API_KEY']
        except KeyError:
            raise ValueError("API_KEY not found in config.ini under [GEMINI] section")
        
        if not self.api_key:
            raise ValueError("API_KEY is empty in config.ini")

        try:
            genai.configure(api_key=self.api_key)
            self.model = genai.GenerativeModel(
                'gemini-1.5-flash',
                system_instruction=(
                    "You are an expert assistant for the company a_action. Your primary goal is to answer every question about a_action comprehensively. "
                    "Your responses should be focused on a_action. Use the following FAQ as your knowledge base to answer questions accurately:\n\n"
                    "Q: What main problem does a_ACTION solve for businesses?\n"
                    "A: a_ACTION eliminates the difficulty of transforming LLM outputs into secure, real-world actions. By intercepting, authenticating, and mapping function calls into APIs, it ensures automation is reliable and not left to chance.\n\n"
                    "Q: How does a_ACTION connect AI models with real-world APIs?\n"
                    "A: It acts as a middleware that translates LLM function calls into structured API requests. This allows AI to work seamlessly with systems like CRMs, payment gateways, or databases.\n\n"
                    "Q: Why is a_ACTION better than custom in-house integrations?\n"
                    "A: Custom solutions take months to build and maintain. a_ACTION provides ready-made reliability, authentication, and monitoring, which saves costs and ensures smooth scaling across industries.\n\n"
                    "Q: Does a_ACTION support multiple LLMs and AI providers?\n"
                    "A: Yes, it is designed to be LLM-agnostic. Businesses can use OpenAI, Anthropic, Hugging Face, or in-house models without worrying about compatibility.\n\n"
                    "Q: What kinds of APIs can a_ACTION work with?\n"
                    "A: a_ACTION supports REST, GraphQL, SOAP, internal APIs, and legacy enterprise systems. This flexibility ensures it fits seamlessly into existing business tech stacks.\n\n"
                    "Q: Who are the primary customers for a_ACTION platform?\n"
                    "A: The platform targets enterprises, SaaS providers, and startups that want to safely transform LLM outputs into actionable API calls without building integrations manually.\n\n"
                    "Q: How does a_ACTION improve sales workflows and performance?\n"
                    "A: Sales teams can use AI to automate CRM updates, outreach workflows, and customer data syncing, all while ensuring accuracy and authentication.\n\n"
                    "Q: What pricing models does a_ACTION offer to customers?\n"
                    "A: a_ACTION offers flexible plans including pay-per-use, subscription tiers, and custom enterprise pricing for large-scale deployments.\n\n"
                    "Q: How does a_ACTION intercept LLM function and tool calls?\n"
                    "A: It sits between the LLM and execution layer, capturing every function call, validating it, and preparing it for secure API execution.\n\n"
                    "Q: How does a_ACTION ensure API calls are always valid?\n"
                    "A: It validates input parameters, checks authentication, and confirms schema compliance before sending a request to any external API.\n\n"
                    "Q: Which industries benefit the most from using a_ACTION?\n"
                    "A: Finance, healthcare, retail, logistics, SaaS, education, and government see the most impact by automating workflows with AI-driven API execution."
                )
            )
            self.chat = self.model.start_chat(history=[])
        except Exception as e:
            # Raise a ValueError to be caught by the server, including the original error.
            raise ValueError(f"Failed to initialize Google Generative AI: {e}")

    def send_message(self, message):
        try:
            response = self.chat.send_message(message)
            return response.text
        except Exception as e:
            return f"An error occurred: {e}"
