a_action_chatbot/requirements.txt
google-generativeai
fastapi
uvicorn
fastapi-cors
